requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
sentence-transformers==2.2.2
faiss-cpu==1.13.0
transformers==4.57.1
# Torch pinned to version compatible with available wheels
torch==2.9.1
python-multipart==0.0.5
aiofiles==23.1.0
requests==2.32.5
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1

